apps/api/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
asyncpg>=0.29.0

# Auth
python-jose[cryptography]>=3.3.0
httpx>=0.27.0

# Validation
pydantic>=2.5.0

# Redis
redis[hiredis]>=5.0.0

# SSE
sse-starlette>=2.0.0

# Config
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
